Refactor loader script to load by folder
Previously I had a file naming scheme to select what to load on each
server, but that didn't worked well because eventually it got mixed with
names I put to differentiate other things beside machines.

So now all things that must be run on a server have their own folder,
which will be compared to `$HOSTNAME` to decide which one to load.

Repeated files con either be moved to `src/common`, which is **always**
loaded. Or link them with symbolic links.