Move around some files
This is part of the move into a way to have a good starting point for
servers as for my machines.

So everything that can go into a server, goes into `src/common`, then
particular stuff to a machine, goes into that machine folder:
`src/$HOST`. Stuff I don't want automatically loaded goes into
`src/other` and can be linked from other places when I want it
automatically loaded on that particular machine.