wezterm: tried to configure starting folder
Didn't work. But I left it for historic purposes. Anyway, after some
more days using `wezterm` I came to realize that's preferable to start
a new tab on the same folder as current tab and do a simple `cd` to go
back $HOME, than to start at $HOME, like in `urxvt` and have to find out
on which folder I want to be.