Add alias for `epub-reader`
I always forget which reader I have installed because throughout the
years I haven't found any that provides me a sane solution for my
reading needs. Most try to become a "library manager", and I just want a
reader. Also most of them are broke or present an awful UI.

`epub-reader` also tries to manage your books, but it looks like a
decent option for reading on the notebook.