Ditch Volta
Volta support for `pnpm` is experimental and even configuring it to work
with `pnpm` doesn't support global packages.

I installed `pnpx` as it didn't came with `pnpm` nor Node.js
installation and it didn't work.

Volta's shim is an ELF binary, so I don't have a clue how to solve this.

Just look for another Node.js version manager and ditch Volta.